[[ Win64 ]] Update engine to compile for win32 x86-64
This patch tweaks the engine code to compile for win32 x86-64.

Mainly it uses intptr_t in the gentle related toolchain code
rather than long, and DWORD_PTR / LONG_PTR appropriately in
Win32 APIs.

Additionally, it updates the build-libs.bat script to use v140
toolset (even if VS 15 is requested - this is because VS2017 does
not have the same structure, and it only bumps the toolset version
by 1, not 10 - i.e. v141 and not v150).

Finally, it modifies fetch-libraries.sh to ensure that the win32
prebuilts have the correct names. The system expects the toolset
to be v141, and for 64-bit it expects the arch to be x64.

+++ toolchain/libcompile/include/position.h
@@ -18,23 +18,24 @@
 #define __POSITION__
 
 #include <stdio.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-typedef long PositionRef;
+typedef intptr_t PositionRef;
 typedef struct File *FileRef;
     
 void InitializePosition(void);
 void FinalizePosition(void);
 
-void AdvanceCurrentPosition(long delta);
+void AdvanceCurrentPosition(intptr_t delta);
 void AdvanceCurrentPositionToNextRow(void);
 void AdvanceCurrentPositionToFile(FileRef file);
 
-void GetColumnOfPosition(PositionRef position, long *r_column);
-void GetRowOfPosition(PositionRef position, long *r_row);
+void GetColumnOfPosition(PositionRef position, intptr_t *r_column);
+void GetRowOfPosition(PositionRef position, intptr_t *r_row);
 void GetFileOfPosition(PositionRef position, FileRef *r_file);
 void GetFilenameOfPosition(PositionRef position, const char **r_filename);
 void GetRowTextOfPosition(PositionRef position, const char **r_text);
@@ -49,9 +50,9 @@ void AddFile(const char *filename);
 int MoveToNextFile(void);
 void GetFilePath(FileRef file, const char **r_path);
 void GetFileName(FileRef file, const char **r_name);
-void GetFileIndex(FileRef file, long *r_index);
-const char *GetFileLineText(FileRef file, long p_row);
-int GetFileWithIndex(long index, FileRef *r_file);
+void GetFileIndex(FileRef file, intptr_t *r_index);
+const char *GetFileLineText(FileRef file, intptr_t p_row);
+int GetFileWithIndex(intptr_t index, FileRef *r_file);
 int GetCurrentFile(FileRef *r_file);
 
 #ifdef __cplusplus
